Fix printf format-string mismatches in NTFS_TRACE* call sites
Latent since NTFS_TRACE* was a no-op (never actually formatted
anything): a union passed where %02X expects an unsigned int,
ULONGLONG
values passed to %u, and a size_t precision arg for %.*ls where printf
expects int. Now that NTFS_TRACE* really calls printf, these would
read misaligned varargs and produce garbage or crash.

+++ src/filename.cpp
@@ -119,7 +119,8 @@ std::wstring_view Filename::GetFilename() const
 
   if (!retval.empty())
   {
-    NTFS_TRACE2("File Name: %.*ls\n", retval.size(), retval.data());
+    NTFS_TRACE2("File Name: %.*ls\n", static_cast<int>(retval.size()),
+                retval.data());
     NTFS_TRACE4("File Permission: %s\t%c%c%c\n",
                 IsDirectory() ? "Directory" : "File", IsReadOnly() ? 'R' : ' ',
                 IsHidden() ? 'H' : ' ', IsSystem() ? 'S' : ' ');
